+++ web/src/main/java/com/homeapp/javatraining/service/UserStatisticsServiceImpl.java
@@ -2,13 +2,11 @@ package com.homeapp.javatraining.service;
 
 import com.homeapp.javatraining.dto.UserTopicStats;
 import com.homeapp.javatraining.model.TestResult;
+import com.homeapp.javatraining.model.Topic;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Slf4j
 @Service
@@ -20,17 +18,19 @@ public class UserStatisticsServiceImpl implements UserStatisticsService {
         log.debug("Calculating user topic statistics, results count={}", results.size());
         Map<String, UserTopicStats> statsByTopic = new HashMap<>();
         for (TestResult result : results) {
-            if (result.getTopic() == null) continue;
-            String displayName = result.getTopic().getDisplayName();
-            UserTopicStats stats = statsByTopic.computeIfAbsent(
-                    displayName,
-                    UserTopicStats::new
-            );
+            Set<Topic> topics = result.getTopics();
+            if (topics == null || topics.isEmpty()) continue;
 
-            stats.incrementTotal();
-
-            if (result.isPassed()) {
-                stats.incrementPassed();
+            for (Topic topic : topics) {
+                String displayName = topic.getDisplayName();
+                UserTopicStats stats = statsByTopic.computeIfAbsent(
+                        displayName,
+                        UserTopicStats::new
+                );
+                stats.incrementTotal();
+                if (result.isPassed()) {
+                    stats.incrementPassed();
+                }
             }
         }